requirements.txt
flask
requests
reportlab
python-nmap
scikit-learn
pyjwt